requirements.txt
# Core PyTorch libraries
torch>=1.12.0
torchvision>=0.13.0

# Data manipulation and CSV handling
pandas>=1.3.0

# Image processing
pillow>=8.3.0

# Visualization (used in inference_visual.py)
matplotlib>=3.5.0